backend/python/excepcional/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.0.7
django-filter==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docopt==0.6.2
drf-yasg==1.17.1
flake8==3.8.3
hypothesis==5.19.0
hypothesis-jsonschema==0.16.1
idna==2.9
inflection==0.4.0
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
multidict==4.7.6
openapi-codec==1.3.2
packaging==20.4
parso==0.7.0
pickleshare==0.7.5
pipreqs==0.4.10
pluggy==0.13.1
pprintpp==0.4.0
prompt-toolkit==3.0.5
py==1.9.0
pycodestyle==2.6.0
pydantic==1.5.1
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
pytest-subtests==0.2.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rich==0.8.13
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
schemathesis==1.1.0
simplejson==3.17.2
six==1.15.0
sortedcontainers==2.2.2
sqlparse==0.3.1
swagger-codegen==0.1.12
toml==0.10.1
traitlets==4.3.3
typer==0.1.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
